+++ src/components/context_menu/index.d.ts
@@ -53,12 +53,22 @@ declare module '@elastic/eui' {
    * @see './context_menu_item.js`
    */
 
-  export type EuiContextMenuItemIcon = ReactElement<any> | string;
+  export type EuiContextMenuItemIcon = ReactElement<any> | string | HTMLElement;
 
-  export interface EuiContextMenuItemProps {
+  export interface EuiContextMenuItemProps extends CommonProps {
     icon?: EuiContextMenuItemIcon;
     hasPanel?: boolean;
+    disabled?: boolean;
+    onClick?: () => void;
     buttonRef?: RefCallback<HTMLButtonElement>;
+    hasPanel?: boolean;
+    toolTipContent?: ReactNode;
+    toolTipTitle?: ReactNode;
+    toolTipPosition?: string;
+    href?: string;
+    target?: string;
+    rel?: string;
+    children?: ReactNode;
   }
 
   export const EuiContextMenuItem: SFC<